Fix warnings for pkb side

+++ Team18/Code18/source/PKB/Storage/PatternStorage.cpp
@@ -14,9 +14,9 @@ bool spa::PatternStorage::isPostfixSubstring(std::string postfix, std::string pa
     return false;
   }
 
-  for (int i = 0; i <= postfixTokens.size() - patternPostfixTokens.size(); ++i) {
+  for (size_t i = 0; i <= postfixTokens.size() - patternPostfixTokens.size(); ++i) {
     bool isMatch = true;
-    for (int j = 0; j < patternPostfixTokens.size(); ++j) {
+    for (size_t j = 0; j < patternPostfixTokens.size(); ++j) {
       if (postfixTokens[i + j] != patternPostfixTokens[j]) {
         isMatch = false;
         break;
